requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
aiohttp==3.4.4
aiomultiprocess==0.4.0
algorithms==0.1.3
altgraph==0.16.1
APScheduler==3.5.3
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
csvkit==1.0.3
cx-Oracle==7.0.0
dbfread==2.0.7
dlib==19.15.0
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
et-xmlfile==1.0.1
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
futures==3.1.1
graphviz==0.9
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
incremental==17.5.0
iobjectspy==9.1.0
ipaddress==1.0.22
isodate==0.6.0
itchat==1.3.10
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
kafka-python==1.4.3
leather==0.3.3
lxml==4.2.5
macholib==1.11
MarkupSafe==1.0
multidict==4.4.0
mutagen==1.41.1
numpy==1.15.1
objgraph==3.4.0
opencv-python==3.4.3.18
openpyxl==2.5.7
pandas==0.23.4
parsedatetime==2.4
parsel==1.5.0
pefile==2018.8.8
Pillow==5.2.0
psutil==5.4.7
py4j==0.10.7
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycrypto==2.7a1
PyDispatcher==2.0.5
pyftpdlib==1.5.4
pygame==1.9.4
PyHamcrest==1.9.0
PyInstaller==3.4
pymongo==3.7.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pypng==0.0.18
Pypubsub==4.0.0
PyQRCode==1.2.1
python-dateutil==2.7.3
python-slugify==1.2.6
pythonnet==2.3.0
pytimeparse==1.1.8
pytz==2018.5
pywin32==223
pywin32-ctypes==0.2.0
queuelib==1.5.0
rarfile==3.0
redis==2.10.6
redis-py-cluster==1.3.5
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
Scrapy==1.5.1
selenium==3.14.0
service-identity==17.0.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.11
threadpool==1.3.2
Twisted==18.7.0
tzlocal==1.5.1
Unidecode==1.0.22
urllib3==1.23
w3lib==1.19.0
Werkzeug==0.14.1
wxpy==0.3.9.8
wxPython==4.0.3
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
yarl==1.2.6
zope.interface==4.5.0
